support header URLs and script interception
Header URLs allow for injecting script URLs at the 
top of HTML, as an alternative or in combination with
previous support for injecting source code as an
inline script.

Script interception allows for intercepting requests
for certain scripts and returning some code string,
rather than forwarding the request to the proxied 
server.